refactor: move dialing utils into iroh from iroh-bytes
this allows iroh-bytes to not have a dependency on iroh-net

rationale: see https://github.com/n0-computer/iroh/issues/1224
fixes https://github.com/n0-computer/iroh/issues/1224
