+++ iroh-bytes/Cargo.toml
@@ -22,7 +22,6 @@ flume = "0.10.14"
 futures = "0.3.25"
 hex = "0.4.3"
 iroh-io = { version = "0.2.1" }
-iroh-net = { path = "../iroh-net" }
 multibase = "0.9.1"
 num_cpus = "1.15.0"
 once_cell = "1.17.0"
